Доработан дизайн элементов списка платежей, изменён дизайн экрана платежей. Добавлены провреки для данных платежа и кнопка выхода из аккаунта.

+++ app/src/main/java/dev/passerby/aeon_project/data/models/PaymentDto.kt
@@ -7,7 +7,7 @@ data class PaymentDto(
     @SerializedName("amount")
     val amount: Any?,
     @SerializedName("created")
-    val created: Int?,
+    val created: Long?,
     @SerializedName("id")
     val id: Int,
     @SerializedName("title")
